[TRAVIS] Try crazy experiment: don't use docker
If I can use the cache for the python dependencies, that removes one of
the main draws for using docker in the first place. I'm going to see
what happens when I remove the container out of the equation. If this
works well, I'll make a separate PR for it